+++ CHANGELOG.md
@@ -9,10 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [0.2.0] - 2019-06-19
+### Added
+* Ability to set target language for translation
+* Support for copying translated text to clipboard
+
+### Changed
+* Translated text is now used as primary text instead of subtext
+
+
 ## [0.1.0] - 2019-06-18
 ### Added
 * Automatic translation of query into English
 
 
-[Unreleased]: https://github.com/dshoreman/albert-translate/compare/v0.1.0...develop
+[Unreleased]: https://github.com/dshoreman/albert-translate/compare/v0.2.0...develop
+[0.2.0]: https://github.com/dshoreman/albert-translate/releases/tag/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/dshoreman/albert-translate/releases/tag/v0.1.0